requirements.txt
Django==1.8.2
celery==3.1.18
djangorestframework==3.0.3
psycopg2==2.5.4
redis==2.10.3
aiohttp==0.16.5
python3-pika==0.9.14
asynqp==0.3